-Model View Controller noch strikter getrennt - Neue FeatureKlassen ( TileTypen) eingefügt - Level nun im Model komplett ansprechbar TODO: - Fehler bei darstellung des levels in der View ( fillLevelWithEntity())

+++ lib/src/model/Blade.dart
@@ -7,7 +7,7 @@ class Blade extends Entity {
   double direction_y = 0.0;
 
   double radius;
-  double spin;
+  int spin;
   bool onfield;
 
   DartBladeGameView view;
